requirements.txt
openai
python-telegram-bot --upgrade
python-dotenv